systemd: prepare for mount_process_proc_self_mountinfo() refactor
When mount_process_proc_self_mountinfo() is refactored to break out the
mount unit state update the sets around and gone will need to be passed
as parameters.

But passing NULL to a function for a set will not see the set initial
creation so pre-allocate the passed in sets.

Doing this before the refactor allows for a no functional change patch
for the refactor that is easier to verify than a compound patch that
does both.